+++ src/timeline-graph/index.js
@@ -1,4 +1,3 @@
 // @flow
-export { TimelineGraph } from "./timeline-graph-DEPRECATED/timeline-graph";
-export { TimelineGraph as CommonTimelineGraph } from "./timeline-graph";
+export { TimelineGraph } from "./timeline-graph";
 export { TimelineCard } from "./timeline-card";